Support C# while loops
While loops are translated by spawning an If chip to test for the supplied condition, and cycling the execution flow of the loop block back to the "entry If" chip for the next iteration.

Break and continue are supported and do what you expect -- break moves execution out of the loop and to the "done" exec flow, and continue moves execution back around to the "entry if" chip.