Keep searching through the shared keyword scope stack until we find a scope that a given keyword has an implementation for
Previously things like using "continue" in a switch statement within a while loop wasn't possible, as we were only checking the shared keyword scope at the top of the stack.

This commit introduces a new method, RunSharedKeywordImpl, that takes a dictionary mapping types to implementations of the keyword.

Additionally, keyword implementations are now passed a reference to the applicable scope.