Wrap blocks in label accessibility scopes during VisitBlock where possible
Ideally we let VisitBlock take care of this most of the time. There are edge cases where it needs to be done manually, though, such as with while/do/if translation, and anonymous single-statement functions (that don't have a block).

The rule of thumb is: if we're creating a block that wasn't there in the original syntax tree, and labels within that block shouldn't be accessible from the parent, we wrap it in a new accessibility scope.