Skip translation of unknown classes
It is hard to reason whether a class should be translated or not. Not translating it seems like the safe approach, where the user has to be specific and add a "CircuitLibrary" class. This doesn't work with inheritance, but if there's a need for it maybe a keyword makes sense?