Add custom chip layout engine
Unity Layouts should be banned from existence. At least from doing bottom-up defined layouts. They don't have a good way to make children control their parents' size.
The workaround is to use content-size-fitters, but every layer of layout depth then needs a full recalc of the layout, which is painful...

The new ChipLayout classes implement a simple layout system that is powerful enough to organize chips in rows and colums.
Some nice side-effects of this are that the formatter is now a lot faster, defined in world units instead of pixels, and also easier to reuse in a standalone application.