Improve circuit formatter
The bfs worked well enough to produce spread out graphs, but it wasn't really easy to read.
This new approach restrucutres the graph into "branches". This makes it easier to follow single lines of executions, as they stay on the same height.

There is still some problems with the formatted graphs. But it will be easier to optimize from this base structure going forward.
The main missing information is a topological grouping, for example to make sure that merge nodes get placed behind all their usage.