Add out keyword to rrcg_return_data in syntax tree
This worked fine before the standalone projects, as the generated code would be re-parsed, but in the syntax tree this refers to an identifier called "out rrcg_return_data", which broke compilation.

By actually adding the out keyword in the syntax tree and preserving the identifier name, we avoid the issue.