Don't add source syntax trees to compilation when rewriting standalone project sources
When it comes time to rewrite the syntax trees, attempting to query the semantic model will, for a currently unknown reason, raise an error that the syntax tree doesn't exist in the compilation.

The current fix is to just add it to the compilation, but if identical syntax trees already exist in the compilation, it leads to ambiguities that cause type resolution to fail, among other things.

For now, at least, adding the syntax trees only when we need to query their semantic information seems to work fine -- but I don't think this is the right way to do it.