Add caching for standalone compilation
Don't blink or you'll miss it!
Also adds a reparsing step after compilation. Rewriting resets the language version (not sure if it can be prevented, it resets to the "Default" version for rewritten syntax trees). This didn't then work in the compilation since files need to have the same language version.