Partially add setter method for __VariableDeclaratorExpression
More investigation shows that we can add the setter method as a parameter, but we need to make sure the variable isn't typed with the "var" keyword.

However this fails for field declarations, where we can't reference the non-static field from the initializer. In this case we'll have to split the field declaration from initialization, and do the latter as part of the static/instance constructor.