Compute variable promotions at rewriting time
This is necessary because for while loops, we want to have the first read of each promoted variable in the block to refer to the variable value port.

However, we only knew if a variable needed to be promoted at assignment time, which means read operations on the variable that occured before assignments would refer to the pre-promotion value.

By looking for assignments to accessible local variables at rewriting time, we can determine which variables need to be promoted ahead of time, solving the issue.