Cleanup conditional context logic
This commit introduces two methods to ConditionalContext: WritePromotedVariables, which writes the current C# value of promoted variables to their CV2 variables, and ResetPromotedVariables, which resets the C# value of promoted variables to the pre-promotion value, or their CV2 variable read pin.

This allows us to do away with the complex branch & final value collection helpers and makes it more immediately obvious what's actually going on.

Additionally, SemanticStack now has a new helper: PopExpectedScope, which replaces a lot of the Pop() comparisons in CircuitBuilder.