Add support for generic RRCG classes
This adds an automatic constraint to AnyPort for generic parameters in a class.
This allows the generic parameter to be used e.g. in variables or generic events.

For now this will always add the constraint. I struggle to think of a use-case where this wouldn't work, but we can re-evaluate if that occurs.