Initial support for conditional returns
In the cases where there are multiple non-void returns, an Event is created on-the-fly that serves as a merge point for the return data. Then, the new event is sent locally at each return site, with the data that was returned.

The old behaviour is retained in the cases where the return type is void/incompatible, or there is only one return.