Add DataTable class chip
This allows users to create Data Tables with RRCG. The constructor takes an enumerable of whole Data Table rows, which allows users to pre-compute the values at build time. The only caveat is that the enumerable must be defined in custom build code (see DataTables test for more on that).