Handle throw statements better across the iterators
In RRCG, a Throw statement is used to end the current branch of execution.

Using a Throw statement in the "main branch" of an iterator is probably something you'd never want to do, but at least we don't crash if you do now.

This also implements a new semantic scope interface - IThrow - which allows you to perform actions before the ExecFlow is cleared.