Manually collect & visit statements in VisitMethodDeclaration
Currently, we avoid inserting AccessibilityScopes in VisitBlock, if the block's parent is one of a few specific syntax kinds. Then, in the visitor for those respective syntax kinds, we insert one there ourselves with the correct kind.

This makes it more convenient for the visitors, but it's not immediately obvious what's going on. A more explicit approach would be to pull the statement visiting up to the visitors, so we avoid VisitBlock entirely.

So this commit is the first step towards doing this, and eliminates the MethodDeclaration case. The following commits will tackle more of these cases, until eventually we can get rid of the parent check altogether.