Fix rewriting fully-qualified source realm types
If the type was referred to by a MemberAccessExpression (e.g. MyNamespace.MyClass), then the rewriting would have happened at the identifier layer (in this case the MyClass node). However, ResolveOrRewriteBuildRealmType prefers to return fully qualified names where it can. So it would have returned the fully-qualified MemberAccessExpression again where Roslyn expects a SimpleName, resulting in an invalid cast crash.